added new class ClickExamples.cs (extracted tests from ApiExamples.cs)
