+++ src/Coypu.AcceptanceTests/Examples/ApiExamples.cs
@@ -109,43 +109,6 @@ namespace Coypu.AcceptanceTests.Examples
         }
 
         [Test]
-        public void Click_example()
-        {
-            var element = Browser.FindButton("clickMeTest");
-            Assert.That(Browser.FindButton("clickMeTest")
-                               .Value,
-                        Is.EqualTo("Click me"));
-
-            element.Click();
-            Assert.That(Browser.FindButton("clickMeTest")
-                               .Value,
-                        Is.EqualTo("Click me - clicked"));
-        }
-
-        [Test]
-        public void ClickButton_example()
-        {
-            Browser.ClickButton("clickMeTest");
-            Assert.That(Browser.FindButton("clickMeTest")
-                               .Value,
-                        Is.EqualTo("Click me - clicked"));
-        }
-
-        [Test]
-        public void ClickLink_example()
-        {
-            Browser.ClickLink("Trigger a confirm");
-            Browser.CancelModalDialog();
-        }
-
-        [Test]
-        public void ClickLinkWithTitle_example()
-        {
-            Browser.ClickLink("Link with title");
-            Browser.CancelModalDialog();
-        }
-
-        [Test]
         public void ConsideringInvisibleElements()
         {
             Browser.FindButton("firstInvisibleInputId", new Options {ConsiderInvisibleElements = true})
